create scroll to top component and add unit test, rename scroll to top button
