+++ src/components/scroll-to-top-button/ScrollToTopButton.js
@@ -1,0 +1,37 @@
+import { useState, useEffect } from 'react'
+
+import IconButton from '@mui/material/IconButton'
+import Box from '@mui/material/Box'
+import ArrowUpwardRoundedIcon from '@mui/icons-material/ArrowUpwardRounded'
+
+import { styles } from '~/components/scroll-to-top-button/ScrollToTopButton.styles'
+
+const ScrollToTopButton = () => {
+  const [isVisible, setIsVisible] = useState(false)
+
+  const scroll = () => (window.scrollY > 450 ? setIsVisible(true) : setIsVisible(false))
+
+  useEffect(() => {
+    window.addEventListener('scroll', scroll)
+    return () => window.removeEventListener('scroll', scroll)
+  }, [])
+
+  const goToTop = () => {
+    window.scrollTo({
+      top: 0,
+      behavior: 'smooth'
+    })
+  }
+
+  return (
+    isVisible && (
+      <Box sx={ styles.root }>
+        <IconButton onClick={ goToTop } sx={ styles.button }>
+          <ArrowUpwardRoundedIcon sx={ styles.icon } />
+        </IconButton>
+      </Box>
+    )
+  )
+}
+
+export default ScrollToTopButton
