+++ src/components/scroll-to-top-button/ScrollToTopButton.styles.js
