+++ src/containers/app-content/AppContent.js
@@ -2,8 +2,9 @@ import Box from '@mui/material/Box'
 
 import AppHeader from '~/containers/layout/app-header/AppHeader'
 import AppMain from '~/containers/layout/app-main/AppMain'
-import ScrollToTop from '~/components/scroll-to-top/ScrollToTop'
+import ScrollToTopButton from '~/components/scroll-to-top-button/ScrollToTopButton'
 import Footer from '~/containers/layout/footer/Footer'
+import ScrollToTop from '~/components/scroll-to-top/ScrollToTop'
 
 const styles = {
   content: {
@@ -25,8 +26,9 @@ const AppContent = () => {
   return (
     <Box data-testid='AppContent' sx={ styles.content }>
       <AppHeader />
-      <AppMain />
       <ScrollToTop />
+      <AppMain />
+      <ScrollToTopButton />
       <Footer />
     </Box>
   )
